Added an extract API for the ruptures [skip hazardlib]
